requirements.txt
django
djangorestframework
django-filters
tabulate